@@ -4,9 +4,10 @@ import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import static android.R.attr.orientation;
 
@@ -21,10 +22,10 @@ public class ListItemDecoration extends RecyclerView.ItemDecoration {
 
     public ListItemDecoration(Context context,int mOrientation) {
         if(orientation != LinearLayoutManager.HORIZONTAL && orientation != LinearLayoutManager.VERTICAL) {
-                         this.mOrientation = DEFAULT_ORIENTATION;
-                     } else {
-                         this.mOrientation = orientation;
-                     }
+            this.mOrientation = DEFAULT_ORIENTATION;
+        } else {
+            this.mOrientation = orientation;
+        }
         m_drawable = context.getResources().getDrawable(R.drawable.divider);
     }
 
@@ -60,7 +61,7 @@ public class ListItemDecoration extends RecyclerView.ItemDecoration {
         int childCount = parent.getChildCount();
         for (int i = 0; i < childCount; i++) {
             View child = parent.getChildAt(i);
-            android.support.v7.widget.RecyclerView v = new android.support.v7.widget.RecyclerView(parent.getContext());
+            RecyclerView v = new RecyclerView(parent.getContext());
             RecyclerView.LayoutParams params = (RecyclerView.LayoutParams) child
                     .getLayoutParams();
             int top = child.getBottom() + params.bottomMargin;
